Only add the sig if they setup the game first

+++ Unverum/UI/MainWindow.xaml.cs
@@ -576,10 +576,13 @@ namespace Unverum
                     {
                         Directory.CreateDirectory(nameWindow.directory);
                         File.Copy(dialog.FileName, $"{nameWindow.directory}{Global.s}{Path.GetFileName(dialog.FileName)}", true);
-                        // Copy over sig if it exists
-                        var sigs = Directory.GetFiles(Path.GetDirectoryName(Global.config.Configs[Global.config.CurrentGame].ModsFolder), "*.sig", SearchOption.TopDirectoryOnly);
-                        if (sigs.Length > 0)
-                            File.Copy(sigs[0], Path.ChangeExtension($"{nameWindow.directory}{Global.s}{Path.GetFileName(dialog.FileName)}", ".sig"), true);
+                        if (Directory.Exists(Global.config.Configs[Global.config.CurrentGame].ModsFolder))
+                        {
+                            // Copy over sig if it exists
+                            var sigs = Directory.GetFiles(Path.GetDirectoryName(Global.config.Configs[Global.config.CurrentGame].ModsFolder), "*.sig", SearchOption.TopDirectoryOnly);
+                            if (sigs.Length > 0)
+                                File.Copy(sigs[0], Path.ChangeExtension($"{nameWindow.directory}{Global.s}{Path.GetFileName(dialog.FileName)}", ".sig"), true);
+                        }
                     }
                 }
             }
